Rename controller/test to controllertest
