+++ sdk/metric/controller/push/push_test.go
@@ -31,8 +31,8 @@ import (
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/export/metric/metrictest"
+	"go.opentelemetry.io/otel/sdk/metric/controller/controllertest"
 	"go.opentelemetry.io/otel/sdk/metric/controller/push"
-	controllerTest "go.opentelemetry.io/otel/sdk/metric/controller/test"
 	"go.opentelemetry.io/otel/sdk/metric/processor/processortest"
 	"go.opentelemetry.io/otel/sdk/resource"
 )
@@ -149,7 +149,7 @@ func TestPushTicker(t *testing.T) {
 	)
 	meter := p.Provider().Meter("name")
 
-	mock := controllerTest.NewMockClock()
+	mock := controllertest.NewMockClock()
 	p.SetClock(mock)
 
 	ctx := context.Background()
@@ -229,7 +229,7 @@ func TestPushExportError(t *testing.T) {
 				push.WithResource(testResource),
 			)
 
-			mock := controllerTest.NewMockClock()
+			mock := controllertest.NewMockClock()
 			p.SetClock(mock)
 
 			ctx := context.Background()
